requirements.txt
aiohttp==3.9.3
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
fastapi-cache2==0.2.1
fonttools==4.50.0
frozendict==2.4.0
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
multidict==6.0.5
multitasking==0.0.11
numpy==1.26.4
orjson==3.9.15
packaging==24.0
pandas==2.2.1
peewee==3.17.1
pendulum==3.0.0
pillow==10.2.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
threadpoolctl==3.3.0
time-machine==2.14.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
yfinance==0.2.37
